requirements.txt
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.2.5
django-annoying==0.10.6
django-crispy-forms==1.7.2
django-filter==2.1.0
django-registration==3.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.9
Markdown==3.1.1
Pillow==5.2.0
PyJWT==1.7.1
pytz==2019.2
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
urllib3==1.25.8
